requirements.txt
fastapi==0.104.1
fastapi-mcp==0.3.0
uvicorn==0.24.0
pydantic==2.5.0
langchain-openai==0.1.8
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
azure-storage-blob==12.19.0
